requirements.txt
python-dotenv
aiogram
aiohttp
asyncpg
sqlalchemy
pydantic
oauthlib
langchain-gigachat
gigachat
alembic
psycopg2
APScheduler